Fix invalid groupby, add SQL export
Two changes in this commite. I think updating Pandas caused issues with
the groupby not working because it was trying to sum a date. In addition
to specifying that the sum should only take place on `time`, I also
specifyied only the required columns to clean up the data being grouped.

An SQL export option was added. In the future, this should be broken out
into its own workflow. There's no need to create graphs if all that is
needed is an SQL file. One option would be to change the arguments to
`watson-graphing graph ...` and `watson-graphing save ...`, although
perhaps a different name would be needed as well.